+++ examples/nextjs-swc/src/appRouterI18n.ts
@@ -35,3 +35,10 @@ export const allI18nInstances: AllI18nInstances = locales.reduce(
   },
   {}
 )
+
+export const getI18nInstance = (locale: SupportedLocales): I18n => {
+  if (!allI18nInstances[locale]) {
+    console.warn(`No i18n instance found for locale "${locale}"`)
+  }
+  return allI18nInstances[locale]! || allI18nInstances['en']!
+}