+++ packages/cli/src/api/catalog/getCatalogs.ts
@@ -132,11 +132,7 @@ export function getCatalogForFile(file: string, catalogs: Catalog[]) {
     const catalogGlob = replacePlaceholders(catalogFile, { locale: "*" })
     const matchPattern = normalizeRelativePath(
       path.relative(catalog.config.rootDir, catalogGlob)
-    )
-      .replace("(", "\\(")
-      .replace(")", "\\)")
-      .replace("[", "\\[")
-      .replace("]", "\\]")
+    ).replace(/(\(|\)|\[|\])/g, "\\$1")
 
     const match = micromatch.capture(matchPattern, normalizeRelativePath(file))
     if (match) {